requirements.txt
pandas
python-telegram-bot
sqlite3

